requirements.txt
aiohttp
jinja2
llama-cpp-python